introduce CFLAGS_APPEND and LDFLAGS_APPEND now used in bitbake recipe
CFLAGS_APPEND and LDFLAGS_APPEND allow to append to CFLAGS and LDFLAGS
instead of overwriting them.  We use EXTRA_OEMAKE to define those and
other make macros we need.  This is in contrast to default value of
EXTRA_OEMAKE which uses make option -e which overwrites makefile
macros with environment variables.